+++ src/App.tsx
@@ -11,7 +11,6 @@ import { CardsPage } from './pages/CardsPage/CardsPage.tsx';
 import { HomePage } from './pages/HomePage/HomePage.tsx';
 import { Navbar } from './components/Navbar/Navbar.tsx';
 import { Footer } from './components/Footer/Footer.tsx';
-import { PokemonPage } from './pages/PokemonPage/PokemonPage.tsx';
 import { WishListPage } from './pages/WishListPage/WishListPage.tsx';
 import { Route, Routes } from 'react-router-dom';
 import './App.scss';
