+++ src/components/SpinLoader/SpinLoader.tsx
@@ -4,7 +4,7 @@ import { Flex, Spin } from 'antd';
 export const SpinLoader = () => {
   return (
     <Flex className="spin-loader" align="center" justify="center">
-      <Spin tip="Loading" size="large" />
+      <Spin size="large" />
     </Flex>
   );
 };
